Fix ESLint errors in Banner, CardLink, and Footer components
