+++ src/components/cardLink/CardLink.scss
@@ -1,7 +1,7 @@
-@import '../styles/variables';
+@import '../../styles/variables';
 
 a.va-api-card {
-  border-top-color: #C4C4C4 !important;
+  border-top-color: #c4c4c4 !important;
 
   &:hover {
     border-top-color: $color-gold !important;
