+++ src/components/cardLink/CardLink.test.tsx
@@ -2,7 +2,7 @@ import { shallow } from 'enzyme';
 import 'jest';
 import * as React from 'react';
 
-import CardLink from './CardLink';
+import { CardLink } from './CardLink';
 
 describe('CardLink', () => {
   it('renders the name', () => {
