+++ src/components/footer/Footer.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import '../../styles/variables';
 
 .va-api-footer-logo {
   width: 20rem;
