+++ src/components/footer/Footer.test.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import 'jest';
 
 import { shallow } from 'enzyme';
-import Footer from './Footer';
+import { Footer } from './Footer';
 
 describe('Footer', () => {
   it('There should be no more beta footer banner', () => {
