+++ src/containers/Home.tsx
@@ -5,7 +5,7 @@ import { Link } from 'react-router-dom';
 import apiDefinitions, { apiCategoryOrder } from '../apiDefs/data/categories';
 import padlockImg from '../assets/homepage-padlock.svg';
 import apiImg from '../assets/homepage-reliable-api.svg';
-import CardLink from '../components/CardLink';
+import { CardLink } from '../components';
 import Hero from '../components/Hero';
 import { Flag } from '../flags';
