+++ src/containers/documentation/CategoryPage.tsx
@@ -3,8 +3,7 @@ import { useParams } from 'react-router';
 import { Flag } from '../../flags';
 import { getApiDefinitions } from '../../apiDefs/query';
 import { APIDescription } from '../../apiDefs/schema';
-import { AuthorizationCard, OnlyTags } from '../../components';
-import CardLink from '../../components/CardLink';
+import { AuthorizationCard, CardLink, OnlyTags } from '../../components';
 import PageHeader from '../../components/PageHeader';
 import { defaultFlexContainer } from '../../styles/vadsUtils';
 import { APINameParam } from '../../types';
@@ -12,7 +11,7 @@ import { PAGE_HEADER_ID } from '../../types/constants';
 
 const CategoryPage = (): JSX.Element => {
   const { apiCategoryKey } = useParams<APINameParam>();
- 
+
   const {
     apis,
     name: categoryName,
