+++ src/containers/releaseNotes/CategoryReleaseNotes.tsx
@@ -6,8 +6,7 @@ import { useParams } from 'react-router';
 import { getDeactivatedCategory, isApiDeactivated } from '../../apiDefs/deprecated';
 import { getApiDefinitions } from '../../apiDefs/query';
 import { APIDescription, BaseAPICategory } from '../../apiDefs/schema';
-import { OnlyTags } from '../../components';
-import CardLink from '../../components/CardLink';
+import { CardLink, OnlyTags } from '../../components';
 import PageHeader from '../../components/PageHeader';
 import { Flag, getFlags } from '../../flags';
 import { defaultFlexContainer } from '../../styles/vadsUtils';
