+++ src/containers/releaseNotes/ReleaseNotesOverview.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { getDeactivatedCategory } from '../../apiDefs/deprecated';
 import { getApiCategoryOrder, getApiDefinitions } from '../../apiDefs/query';
-import CardLink from '../../components/CardLink';
+import { CardLink } from '../../components';
 import PageHeader from '../../components/PageHeader';
 import { Flag } from '../../flags';
 import { defaultFlexContainer } from '../../styles/vadsUtils';
